+++ src/hooks/appdelegate.hpp
@@ -3,6 +3,9 @@
 #include <Geode/modify/AppDelegate.hpp>
 
 class $modify(GlobedAppDelegate, AppDelegate) {
+    $override
     void applicationDidEnterBackground();
+
+    $override
     void applicationWillEnterForeground();
 };