+++ src/hooks/pause_layer.hpp
@@ -3,5 +3,6 @@
 #include <Geode/modify/PauseLayer.hpp>
 
 class $modify(GlobedPauseLayer, PauseLayer) {
+    $override
     void customSetup();
 };