+++ src/hooks/player_object.hpp
@@ -10,11 +10,15 @@ class $modify(ComplexPlayerObject, PlayerObject) {
     void setRemoteState();
     void setDeathEffect(int deathEffect);
 
+    $override
     void incrementJumps();
+
+    $override
     void playDeathEffect();
 };
 
 // Unlike `ComplexPlayerObject`, this one is made specifically for vanilla player objects, so it is a separate $modify class.
 class $modify(HookedPlayerObject, PlayerObject) {
+    $override
     void playSpiderDashEffect(cocos2d::CCPoint from, cocos2d::CCPoint to);
 };